# Add benchmarks and optimize CanConvert