Significant improvement for #12.

Benchmark results:

Before:

```
 |                         Method |       Mean |    StdErr |    StdDev |     Median |
 |------------------------------- |----------- |---------- |---------- |----------- |
 |         CanConvert_Int32_Int32 | 32.3227 ns | 0.1344 ns | 0.5205 ns | 32.2007 ns |
 | CanConvert_Int32_NullableInt32 | 51.1027 ns | 0.3238 ns | 1.2117 ns | 51.0040 ns |
 |        CanConvert_Int32_Object | 49.7077 ns | 0.5168 ns | 2.1924 ns | 49.0883 ns |
 |        CanConvert_Int32_String | 51.6088 ns | 0.4570 ns | 1.7101 ns | 51.0255 ns |
 |        CanConvert_Int32_UInt32 | 53.4829 ns | 0.5534 ns | 3.4115 ns | 53.3924 ns |
 |       CanConvert_String_String | 34.2702 ns | 0.3612 ns | 2.0750 ns | 33.3913 ns |
 |       CanConvert_String_Object | 50.7273 ns | 0.4690 ns | 1.7547 ns | 50.2128 ns |
 |       CanConvert_String_UInt32 | 51.9694 ns | 0.2787 ns | 1.0427 ns | 52.1386 ns |
 |       CanConvert_Stream_Stream | 34.1380 ns | 0.0557 ns | 0.1929 ns | 34.1391 ns |
 | CanConvert_Stream_MemoryStream | 49.8007 ns | 0.5182 ns | 4.0472 ns | 49.0353 ns |
 |       CanConvert_Stream_Object | 52.3670 ns | 0.5382 ns | 3.9549 ns | 50.5785 ns |
 |       CanConvert_Stream_String | 50.1581 ns | 0.2479 ns | 0.9274 ns | 49.7345 ns |
 |       CanConvert_Stream_UInt32 | 51.7420 ns | 0.3143 ns | 1.0888 ns | 51.7160 ns |
```

After:

```
 |                         Method |       Mean |    StdDev |
 |------------------------------- |----------- |---------- |
 |         CanConvert_Int32_Int32 |  3.7360 ns | 0.0923 ns |
 | CanConvert_Int32_NullableInt32 |  7.2751 ns | 0.0991 ns |
 |        CanConvert_Int32_Object |  7.5882 ns | 0.2187 ns |
 |        CanConvert_Int32_String |  8.0872 ns | 0.2671 ns |
 |        CanConvert_Int32_UInt32 |  7.6310 ns | 0.2044 ns |
 |       CanConvert_String_String |  4.7130 ns | 0.0883 ns |
 |       CanConvert_String_Object | 15.9726 ns | 0.4164 ns |
 |       CanConvert_String_UInt32 | 15.1037 ns | 0.1784 ns |
 |       CanConvert_Stream_Stream |  4.6459 ns | 0.1040 ns |
 | CanConvert_Stream_MemoryStream | 52.7746 ns | 1.0389 ns |
 |       CanConvert_Stream_Object | 52.0123 ns | 0.8069 ns |
 |       CanConvert_Stream_String | 53.4943 ns | 1.4109 ns |
 |       CanConvert_Stream_UInt32 | 55.4518 ns | 1.6362 ns |
```

Out of our 11 converters:

- 9 are for value types (LocalDate, LocalDateTime, LocalTime, Instant, Offset, OffsetDateTime, ZonedDateTime, Offset, Duration, Interval)
- 1 is for a sealed class (Period)
- 1 is for an unsealed class (DateTimeZone) - this is the one that wouldn't see any improvement, yet at least

So I expect this to have a pretty significant impact.

@willdean: How easily would you be able to test this, either before or after a merge?