@willdean 2017-06-08
@jskeet I haven't got time to make a proper PR this AM, but here's my bench-mark code:

https://gist.github.com/willdean/11ce3cac28cf829ebec9784e5598401b