@jskeet 2017-06-08
No problems in terms of a PR - I'll add that code to this one, assuming you're happy with that.
(I'll add it to the first commit, then add the benchmark results before/after here.)