@willdean 2017-06-08
@jskeet Sure, no problem - use it however it helps.

I am just about to try this PR-branch binary on my system here to get a direct side-by-side