@willdean 2017-06-08
It seems to be better, but still bad.   I will profile it next.

```
Old                           
                           
                           Method |       Mean |    Error |   StdDev | Scaled | ScaledSD |  Gen 0 | Allocated |
--------------------------------- |-----------:|---------:|---------:|-------:|---------:|-------:|----------:|
 SerializeDateTimeWithoutSettings |   956.9 ns | 18.93 ns | 20.25 ns |   1.00 |     0.00 | 0.3605 |   1.11 KB |
  SerializeInstantWithOutSettings | 1,714.4 ns | 21.47 ns | 20.09 ns |   1.79 |     0.04 | 0.5093 |   1.57 KB |
    SerializeDateTimeWithSettings | 3,286.5 ns | 63.47 ns | 91.02 ns |   3.44 |     0.12 | 0.4196 |    1.3 KB |
     SerializeInstantWithSettings | 2,855.7 ns | 60.86 ns | 70.08 ns |   2.99 |     0.09 | 0.5226 |   1.62 KB |
     

New   
     
                           Method |       Mean |    Error |   StdDev | Scaled | ScaledSD |  Gen 0 | Allocated |
--------------------------------- |-----------:|---------:|---------:|-------:|---------:|-------:|----------:|
 SerializeDateTimeWithoutSettings |   944.0 ns | 21.57 ns | 23.98 ns |   1.00 |     0.00 | 0.3614 |   1.11 KB |
  SerializeInstantWithOutSettings | 1,710.0 ns | 14.96 ns | 13.26 ns |   1.81 |     0.05 | 0.5093 |   1.57 KB |
    SerializeDateTimeWithSettings | 2,248.5 ns | 38.76 ns | 34.36 ns |   2.38 |     0.07 | 0.4196 |    1.3 KB |
     SerializeInstantWithSettings | 2,209.5 ns | 27.25 ns | 24.16 ns |   2.34 |     0.06 | 0.5226 |   1.62 KB |     
     
```