@jskeet 2017-06-08
My results:

Before:

```
 |                             Method |          Mean |     StdDev | Scaled | Scaled-StdDev |
 |----------------------------------- |-------------- |----------- |------- |-------------- |
 |  SerializeDateTime_DefaultSettings |   705.3737 ns |  3.8530 ns |   1.00 |          0.00 |
 |   SerializeInstant_DefaultSettings |   799.0672 ns | 48.9597 ns |   1.13 |          0.07 |
 | SerializeDateTime_NodaTimeSettings | 2,869.5772 ns | 16.6888 ns |   4.07 |          0.03 |
 |  SerializeInstant_NodaTimeSettings | 2,584.6233 ns | 57.1635 ns |   3.66 |          0.08 |
```

After:

```
 |                             Method |          Mean |     StdDev | Scaled | Scaled-StdDev |
 |----------------------------------- |-------------- |----------- |------- |-------------- |
 |  SerializeDateTime_DefaultSettings |   700.3885 ns | 19.0739 ns |   1.00 |          0.00 |
 |   SerializeInstant_DefaultSettings |   767.2527 ns | 17.2677 ns |   1.10 |          0.04 |
 | SerializeDateTime_NodaTimeSettings | 1,854.3514 ns | 41.9430 ns |   2.65 |          0.09 |
 |  SerializeInstant_NodaTimeSettings | 1,976.8734 ns | 33.8632 ns |   2.82 |          0.09 |
```

So yes, a pretty big improvement, but more could be done.

I have two other ideas:

- Change `CanConvert` to first check whether the namespace is `NodaTime`. Really, really hacky...
- Make a "super converter" so that only one converter is added, and delegates to the others where appropriate. Not a big fan of that to be honest, but it would stop there being quite as many calls...