@jskeet 2017-06-08
Just tested, and adding *just* the `Instant` converter reduces `SerializeDateTime_NodaTimeSettings` to 838ns, which is pretty close to the default settings version.

So if we can work out a way of delegating through a single converter that can do a first "is this Noda Time?" pass really quickly, that would help. However, we probably wouldn't want to change the existing extension method to do that, as it'll be a subtle but visible change - users may be replacing individual converters, in which case the single delegating converter would break that. Given that, I suggest we still merge this change.