@willdean 2017-06-08
@jskeet I am still playing here, because it looks like having *any* converters which are slow is a disaster, so I forced `CheckAssignableFrom` to false without as much improvement as I expected.

Obviously I have had to radically change my entire setup in the middle of this as I've changed from using your binaries via Nuget to having everything in on VS solution building here.

Further to your comment about the super-converter which you've just added, I suspect the problem is node adding 10 converters, it's adding nine fast ones and 1 slow one.

I have to go and vote now, back in a bit...