@jskeet 2017-06-08
The only converter which has any business being slow is the one for `DateTimeZone` - and removing that only has a small-ish impact. Removing half the converters reduces the damage by "about half", so I think it really is a matter of reducing that as far as possible... or trying to contribute a patch to Json.NET to avoid it from calling `CanConvert` quite so much.