@jskeet 2017-06-08
Note to self: removing the `CheckAssignableFrom` part entirely doesn't make much difference.
I'm going to file an issue with Json.NET to see whether James has any suggestions.