@willdean 2017-06-08
@jskeet I think talking to JNK is a good idea - I agree it's a bit hard to see where all the time is going.  You would think that all the string conversion stuff would be *way* more expensive than this.  I shall play a bit more when I can.