@willdean 2017-06-08
As a sort of *reductio ad absurdum* I reduced `CanConvert` to just `public override bool CanConvert(Type objectType) => false ;` and obtained these results:

```
                           Method |       Mean |    Error |   StdDev | Scaled |
--------------------------------- |-----------:|---------:|---------:|-------:|
 SerializeDateTimeWithoutSettings |   853.3 ns | 3.589 ns | 3.182 ns |   1.00 |
    SerializeDateTimeWithSettings | 1,997.5 ns | 5.404 ns | 4.219 ns |   2.34 |
```

Wherever the solution lies, it's clear that there's no version of `CanConvert` which will recover the lost performance.

I am wondering if there are things involving `DataContracts` which might be better cached with json.net