@jskeet 2017-06-14
Just going through this. I think I'd prefer to keep this to just one place rather than have it in every subclass... it still needs to behave the same way even if subclasses override as a matter of optimization, and I don't think that optimization achieves much anyway.

As such, I think I'd like to merge as-is for now. @malcolmr: Were you suggesting any changes in the tests for the current code, or only if we put logic into subclasses?