@malcolmr 2017-06-15
@jskeet test changes only if we pushed the logic into subclasses. As-is is fine.